+++ GUI_20212202_IJA9WQ/Models/AbstractGameItem.cs
@@ -9,10 +9,10 @@ namespace GUI_20212202_IJA9WQ.Models
 {
     public abstract class GameItem
     {
-        protected int placeX;
-        protected int placeY;
-        protected int displayWidth;
-        protected int displayHeight;
+        protected double placeX;
+        protected double placeY;
+        protected double displayWidth;
+        protected double displayHeight;
         public abstract Geometry Area { get; }
         public abstract void Terminated();
